refactor(marketing-analytics): name the sections file like the rest of the frontend
SETUP_SECTIONS.tsx was the only SCREAMING_SNAKE filename in frontend/src —
the constant it exports is spelled that way, the file shouldn't be. camelCase
matches its own sibling sectionRouting.ts and the activityDescriptions.tsx
precedent for a JSX-carrying module that isn't a component.

Generated-By: PostHog Desktop
Task-Id: ecf01baa-6eaf-43e8-b9ce-ae4c0860add4
